Implemented ability for user to set and retrieve custom bitrates for all formats and resolutions, video and live. Refactored deleteDirectoryRecursive to be best-effort. Refactored live stream image generation and processing, also fixing memory leak.

+++ utils/helpers.js
@@ -68,23 +68,37 @@ function logDebugMessageToConsole(message, error, stackTrace, isLoggingToFile) {
 }
 
 function deleteDirectoryRecursive(directoryPath) {
-    const fs = require('fs');
-    const path = require('path');
+	try {
+		if(fs.existsSync(directoryPath)) {
+			fs.readdirSync(directoryPath).forEach((file) => {
+				try {
+					const currentPath = path.join(directoryPath, file);
 
-    if (fs.existsSync(directoryPath)) {
-        fs.readdirSync(directoryPath).forEach((file) => {
-            const curPath = path.join(directoryPath, file);
-            
-            if (fs.statSync(curPath).isDirectory()) {
-                deleteDirectoryRecursive(curPath);
-            }
-            else {
-                fs.unlinkSync(curPath);
-            }
-        });
-        
-        fs.rmdirSync(directoryPath);
-    }
+					if (fs.statSync(currentPath).isDirectory()) {
+						deleteDirectoryRecursive(currentPath);
+					}
+					else {
+						fs.unlinkSync(currentPath);
+					}
+				}
+				catch(e) {
+					// do nothing
+				}
+			});
+
+			try {
+				if (fs.readdirSync(directoryPath).length === 0) {
+					fs.rmdirSync(directoryPath);
+				}
+			}
+			catch(e) {
+				// do nothing
+			}
+		}
+	}
+	catch(e) {
+		// do nothing
+	}
 }
 
 function timestampToSeconds(timestamp) {
